requirements.txt
streamlit
requests
pyperclip
python-dotenv
google-generative-ai